Add enums package
